# public/openapi.yaml
servers:
- url: http://127.0.0.1:5000/
  description: The development API server
security:
- ApiKeyAuth: []
paths:
  /login: {}
  /logout: {}
  /self-update: {}
  /tasks:
    get:
      summary: Get all tasks
      responses:
        '200':
          description: Returns all tasks with their statuses
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    detail:
                      type: string
                    done:
                      type: boolean
    post:
      summary: Add or complete tasks
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                add:
                  type: array
                  items:
                    type: string
                complete:
                  type: array
                  items:
                    type: string
      responses:
        '200':
          description: Returns all tasks with their statuses
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    detail:
                      type: string
                    done:
                      type: boolean
  /api/docs/dist/{filename}: {}
  /api/docs/{path}: {}
  /: {}
  /openapi.yaml: {}
  /app-list:
    get:
      summary: List all Chat Dokku apps
      description: Returns a list of all Chat Dokku apps.
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                type: array
                items:
                  type: string
  /app-create:
    get:
      summary: Create a new web app
      description: Create a new web app, defaulting to a unique random name
      parameters:
      - name: app_name
        in: query
        description: Name of the app to create
        required: false
        schema:
          type: string
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ExecResult'
  /write-file:
    post:
      summary: Write a file
      description: Write a file
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                app_name:
                  type: string
                  description: Name of the app where the file will be written
                path:
                  type: string
                  description: Path inside the app where the file will be written
                contents:
                  type: string
                  description: Contents of the file to be written
              required:
              - app_name
              - path
              - contents
  /exec-in-workdir:
    post:
      summary: Write a file
      description: Write a file
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                app_name:
                  type: string
                  description: Name of the app where the command will be executed
                command:
                  type: string
                  description: Shell command to run
              required:
              - app_name
              - command
info:
  title: Chat Dokku
  version: 1.0.0
openapi: 3.0.1
components:
  schemas:
    ExecResult:
      type: object
      properties:
        returncode:
          type: integer
        output:
          type: string
      required:
      - output
      - returncode
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: Authorization